fix: correcting sorting by date
Create separate DatePickerModal component to fix navigation.
