+++ components/DatePickerModal/type.ts
@@ -1,0 +1,4 @@
+export interface DatePickerModalProps {
+    show: boolean;
+    setShow: (value: boolean) => void;
+}
