+++ utils/formatDate.ts
@@ -13,5 +13,5 @@ export const formatDateWithoutTime = (date: Date) => {
   const month = String(date.getMonth() + 1).padStart(2, '0');
   const year = date.getFullYear();
 
-  return `${day}.${month}.${year}`;
+  return `${year}-${month}-${day}`;
 };
